add tests for changes
